Add no items composable for list
