+++ app/src/main/java/com/slaviboy/drumpadmachine/screens/presets/composables/PresetsComposable.kt
@@ -3,20 +3,15 @@ package com.slaviboy.drumpadmachine.screens.presets.composables
 import androidx.compose.animation.core.TargetBasedAnimation
 import androidx.compose.animation.core.VectorConverter
 import androidx.compose.animation.core.tween
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
@@ -28,38 +23,23 @@ import androidx.compose.runtime.withFrameNanos
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.onGloballyPositioned
-import androidx.compose.ui.layout.positionInRoot
 import androidx.compose.ui.platform.LocalSoftwareKeyboardController
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.TextUnit
-import com.bumptech.glide.integration.compose.CrossFade
-import com.bumptech.glide.integration.compose.ExperimentalGlideComposeApi
-import com.bumptech.glide.integration.compose.GlideImage
-import com.bumptech.glide.integration.compose.placeholder
 import com.ramcosta.composedestinations.annotation.Destination
 import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.slaviboy.composeunits.dh
 import com.slaviboy.composeunits.dw
 import com.slaviboy.composeunits.sw
 import com.slaviboy.drumpadmachine.R
+import com.slaviboy.drumpadmachine.composables.NoItems
 import com.slaviboy.drumpadmachine.composables.ScrollableContainer
 import com.slaviboy.drumpadmachine.data.entities.Preset
 import com.slaviboy.drumpadmachine.events.ErrorEvent
 import com.slaviboy.drumpadmachine.events.NavigationEvent
 import com.slaviboy.drumpadmachine.extensions.ObserveAsEvents
-import com.slaviboy.drumpadmachine.extensions.bounceClick
 import com.slaviboy.drumpadmachine.extensions.mapValue
-import com.slaviboy.drumpadmachine.modules.NetworkModule
 import com.slaviboy.drumpadmachine.screens.destinations.DrumPadComposableDestination
 import com.slaviboy.drumpadmachine.screens.home.composables.HomePresetDetails
 import com.slaviboy.drumpadmachine.screens.presets.viewmodels.PresetsViewModel
-import com.slaviboy.drumpadmachine.ui.RobotoFont
 import kotlin.math.roundToInt
 
 @OptIn(ExperimentalComposeUiApi::class)
@@ -133,6 +113,9 @@ fun PresetsComposable(
             targetValue = 1f
         )
     }
+    var topBarHeight by remember {
+        mutableStateOf(0.dw)
+    }
     LaunchedEffect(animationFlag) {
         animationFlag ?: return@LaunchedEffect
         val startTime = withFrameNanos { it }
@@ -168,6 +151,7 @@ fun PresetsComposable(
         minHeight = 0.36.dw,
         maxHeight = 0.53.dw,
         topBar = { height, minHeight, maxHeight ->
+            topBarHeight = height
             PresetsTopBar(
                 title = name,
                 subtitle = "Search for your favorite sound pack",
@@ -277,81 +261,19 @@ fun PresetsComposable(
                 animationFlag = !(animationFlag ?: true)
             }
         )
-    }
-}
 
-@OptIn(ExperimentalGlideComposeApi::class)
-@Composable
-fun PresetCard(
-    preset: Preset,
-    titleTextSize: TextUnit,
-    subtitleTextSize: TextUnit,
-    coverSize: Dp,
-    modifier: Modifier = Modifier,
-    onPresetClick: (x: Float, y: Float, preset: Preset) -> Unit
-) {
-    var x by remember {
-        mutableFloatStateOf(0f)
-    }
-    var y by remember {
-        mutableFloatStateOf(0f)
-    }
-    Column(
-        modifier = modifier
-            .width(coverSize)
-            .onGloballyPositioned {
-                val position = it.positionInRoot()
-                x = position.x
-                y = position.y
-            }
-            .bounceClick {
-                onPresetClick(x, y, preset)
-            }
-    ) {
-        Box(
-            contentAlignment = Alignment.BottomEnd
-        ) {
-            GlideImage(
-                model = NetworkModule.coverIconUrl(preset.id),
-                contentDescription = null,
+        presetsViewModel.noItemsState.value?.let {
+            NoItems(
+                boxScope = this,
                 modifier = Modifier
-                    .size(coverSize)
-                    .clip(RoundedCornerShape(0.04.dw)),
-                transition = CrossFade,
-                failure = placeholder(R.drawable.ic_no_image)
-            )
-            Image(
-                painter = painterResource(id = R.drawable.ic_play_button),
-                contentDescription = null,
-                modifier = Modifier
-                    .size(coverSize * 0.18f)
-                    .offset(
-                        x = (-0.02).dw,
-                        y = (-0.02).dw
-                    )
+                    .padding(
+                        top = topBarHeight,
+                        bottom = 0.075.dh
+                    ),
+                iconResId = it.iconResId,
+                titleResId = R.string.no_items,
+                subtitleResId = R.string.please_check_your_network
             )
         }
-        Spacer(
-            modifier = Modifier
-                .height(0.02.dw)
-        )
-        Text(
-            text = preset.name,
-            color = Color.White,
-            fontFamily = RobotoFont,
-            fontSize = titleTextSize,
-            fontWeight = FontWeight.Bold,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis
-        )
-        Text(
-            text = preset.author ?: "",
-            color = Color.LightGray,
-            fontFamily = RobotoFont,
-            fontSize = subtitleTextSize,
-            fontWeight = FontWeight.Normal,
-            maxLines = 1,
-            overflow = TextOverflow.Ellipsis
-        )
     }
 }
