+++ app/src/main/java/com/slaviboy/drumpadmachine/screens/presets/viewmodels/PresetsViewModel.kt
@@ -5,12 +5,14 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.slaviboy.drumpadmachine.R
 import com.slaviboy.drumpadmachine.api.results.Result
 import com.slaviboy.drumpadmachine.data.entities.Preset
 import com.slaviboy.drumpadmachine.events.ErrorEvent
 import com.slaviboy.drumpadmachine.events.NavigationEvent
 import com.slaviboy.drumpadmachine.extensions.containsString
 import com.slaviboy.drumpadmachine.screens.home.usecases.DownloadAudioZipUseCase
+import com.slaviboy.drumpadmachine.screens.home.viewmodels.BaseItem
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.receiveAsFlow
@@ -35,23 +37,12 @@ class PresetsViewModel @Inject constructor(
     private val errorEventChannel = Channel<ErrorEvent>()
     val errorEventFlow = errorEventChannel.receiveAsFlow()
 
+    private val _noItemsState: MutableState<BaseItem?> = mutableStateOf(null)
+    val noItemsState: State<BaseItem?> = _noItemsState
+
     fun changeText(text: String) {
         _searchTextState.value = text
         search()
-    }
-
-    fun search() = viewModelScope.launch {
-        if (_searchTextState.value.isEmpty()) {
-            _filteredPresetsState.value = _presetsState.value
-            return@launch
-        }
-        val list = mutableListOf<Preset>()
-        _presetsState.value.forEach {
-            if (it.name.containsString(_searchTextState.value)) {
-                list.add(it)
-            }
-        }
-        _filteredPresetsState.value = list
     }
 
     fun initPresets(presets: Array<Preset>) {
@@ -85,6 +76,34 @@ class PresetsViewModel @Inject constructor(
     }
 
     fun unlockAllSounds() {
+    }
+
+    private fun search() = viewModelScope.launch {
+        if (_searchTextState.value.isEmpty()) {
+            _filteredPresetsState.value = _presetsState.value
+            return@launch
+        }
+        val list = mutableListOf<Preset>()
+        _presetsState.value.forEach {
+            if (it.name.containsString(_searchTextState.value)) {
+                list.add(it)
+            }
+        }
+        _filteredPresetsState.value = list
+        setNoItemEvent()
+    }
+
+    private fun setNoItemEvent() {
+        val isEmpty = _filteredPresetsState.value.isEmpty()
+        _noItemsState.value = if (isEmpty) {
+            BaseItem(
+                iconResId = R.drawable.ic_audio,
+                titleResId = R.string.no_items,
+                subtitleResId = R.string.there_are_no_items_found
+            )
+        } else {
+            null
+        }
     }
 
     private fun getPresetById(presetId: Int): Preset? {
